notifications_enabled = 0 in both regions
The previous fix to change environment variables from overrides to
defaults introduced a new bug: notifications_enabled was being set to 0
in both regions. This fix uses a default value of 2 in puppet to signal
puppetdb-stencil to use the environment variable as an override. But if
puppet sets notifications_enabled to 0 or 1, it will be used and the
environment variable will be ignored.

Issue: CIRRUS-4549